use taskkill on windows

+++ electron/main.development.js
@@ -74,12 +74,21 @@ if (isEtcApi && mantisCmd && mantisPath) {
 
   // Stop Mantis (on app quit)
   app.on('will-quit', () => {
-    Log.info('Stopping Mantis...');
-    psTree(mantis.pid, (err, children) => {
-      // Kill all Mantis child processes
-      children.forEach(function (proc) { process.kill(proc.PID); });
-    });
-    process.kill(mantis.pid); // Kill main Mantis process
+    Log.info('Stopping Mantis(PID '+mantis.pid+')...');
+    if (process.platform == 'win32') {
+      Log.info("with taskkill");
+      spawn("taskkill", ["/F", "/T", "/PID", mantis.pid]); // Kill main Mantis process
+    } else {
+      Log.info("with process.kill");
+      psTree(mantis.pid, (err, children) => {
+        // Kill all Mantis child processes
+        children.forEach(function (proc) {
+          Log.info("and child " + proc.PID");
+          process.kill(proc.PID);
+        });
+      });
+      process.kill(mantis.pid); // Kill main Mantis process
+    }
   });
 }
